toy-les/requirements.txt
numpy>=1.24
torch>=2.1
PyYAML>=6.0
matplotlib>=3.8
